Restructured UTs of rest modules of data-access-layer
