+++ broker/data-access-layer/docker/test/docker.DockerPortRegistry.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const proxyquire = require('proxyquire');
-const DockerPortRegistry = proxyquire('../../data-access-layer/docker/src/DockerPortRegistry', {
+const DockerPortRegistry = proxyquire('../src/DockerPortRegistry', {
   lodash: {
     sample: function (collection) {
       return collection[2];
